Extracted constants used in managing the calendar, such as those changing calendar view, and requests to go to the next page, previous page and today's page. Constants are placed in CalendarFxUtil.
