+++ src/main/java/seedu/address/logic/commands/ViewNextCommand.java
@@ -1,5 +1,7 @@
 package seedu.address.logic.commands;
 //@@author SuxianAlicia
+import static seedu.address.ui.util.CalendarFxUtil.REQUEST_NEXT;
+
 import seedu.address.commons.core.EventsCenter;
 import seedu.address.commons.events.ui.ChangeCalendarPageRequestEvent;
 import seedu.address.logic.commands.exceptions.CommandException;
@@ -21,7 +23,6 @@ public class ViewNextCommand extends Command {
             + "Example: " + COMMAND_WORD;
 
     public static final String MESSAGE_VIEW_CALENDAR_NEXT_SUCCESS = "Displayed next page in Calendar.";
-    public static final String REQUEST_NEXT = "Next";
 
     @Override
     public CommandResult execute() throws CommandException {
