+++ src/main/java/seedu/address/ui/CalendarPanel.java
@@ -1,6 +1,12 @@
 package seedu.address.ui;
 //@@author SuxianAlicia
 import static java.util.Objects.requireNonNull;
+import static seedu.address.ui.util.CalendarFxUtil.DAY_VIEW;
+import static seedu.address.ui.util.CalendarFxUtil.MONTH_VIEW;
+import static seedu.address.ui.util.CalendarFxUtil.REQUEST_BACK;
+import static seedu.address.ui.util.CalendarFxUtil.REQUEST_NEXT;
+import static seedu.address.ui.util.CalendarFxUtil.REQUEST_TODAY;
+import static seedu.address.ui.util.CalendarFxUtil.WEEK_VIEW;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
@@ -26,14 +32,6 @@ import seedu.address.ui.util.CalendarFxUtil;
  * ContactSails implements CalendarFX to display Calendar.
  */
 public class CalendarPanel extends UiPart<Region> {
-
-    private static final String DAY_VIEW = "Day";
-    private static final String MONTH_VIEW = "Month";
-    private static final String WEEK_VIEW = "Week";
-
-    private static final String REQUEST_TODAY = "Today";
-    private static final String REQUEST_BACK = "Back";
-    private static final String REQUEST_NEXT = "Next";
 
     private static final String FXML = "CalendarPanel.fxml";
