+++ src/test/java/seedu/address/logic/commands/ViewBackCommandTest.java
@@ -3,6 +3,7 @@ package seedu.address.logic.commands;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
+import static seedu.address.ui.util.CalendarFxUtil.REQUEST_BACK;
 
 import org.junit.Rule;
 import org.junit.Test;
@@ -56,6 +57,6 @@ public class ViewBackCommandTest {
 
         ChangeCalendarPageRequestEvent lastEvent = (ChangeCalendarPageRequestEvent) eventsCollectorRule.eventsCollector
                 .getMostRecent();
-        assertEquals(ViewBackCommand.REQUEST_BACK, lastEvent.getRequestType());
+        assertEquals(REQUEST_BACK, lastEvent.getRequestType());
     }
 }
